Optimized GIGraph rendering on repos with huge amount of branches
NSMutableArray usage in [GIGraph _generateGraph] caused a bottleneck due to a lot of calls
to the [NSMutableArray containsObject:] and [NSMutableArray removeObject:] methods.
These methods are really slow on a huge amount of tips (~24000 in my case).

To solve this issue CGOrderedSet class was created:
* optimized for speed of adding/removal objects;
* preserving the objects order.

NSMutableArray was replaced to GCOrderedSet which solved this particular bottleneck.