Refactor GitUpKit Base classes into separate static library
The GitUpKit Core and Extension sources are compiled into a static library named libGitUpKitBase and the XLFacility sources are compiled into a static library named libXLFacility. This allows for more modular dependency management in other targets (Tests, GitUpKit Framework). It also makes the base GitUpKit functionality available for use on iOS in the future, since it only depends on Foundation.