Always show welcome when clicking on dock icon
Before this change, if the (x) button was clicked to close the welcome
window, then clicking on the app icon in the dock did nothing and there
was no way to show the welcome window again. Now, clicking on the dock
icon will show the welcome window if there are no other visible windows.

This is how Xcode behaves, which seems like a good standard to emulate
and something many GitUp users would expect.