Store map view options in user defaults
Use the repository-specific settings if set, then fall back to the user defaults. This means when opening a repository for the first time it will use the last settings that were set in any other repository. This is good as I almost always want the same settings in all repositories.