Use NSFileManager instead of exchangedata for index updates
The high level API deals with file metadata.

Fixes exchangedata not being available with APFS in this case.

The only bit I’m not 100% about is the role of the line by ‘Touch file to make sure any cached information in the index gets invalidated’. I’ve been using this for a while and not noticed any problems.