Ignore block capturing autoreleased objects warnings
This warning is raised in several places.

I think added NS_NOESCAPE ought to fix most of these, but it does not. That seems to be a compiler limitation.

They could also be avoiding by making local variables, which would also avoid the casts from (for example) NSArray to NSMutableArray. However making these changes requires care to avoid breaking the goto statements.