Allow title bar double-clicks to zoom window
Standard macOS behavior when the user double-clicks the title bar of a window is to zoom/unzoom the window. The user can also choose, in System Preferences > Dock, to minimize the window on a title-bar double-click.

AppKit's implementation of the unified title/toolbar (which GitUp uses) breaks this behavior. This commit works around AppKit's bad behavior. See -[ToolbarItemWrapperView hitTest:] for details.