Safely reopen repository when it is renamed instead of possibly crashing
This addresses a crash that became consistent on macOS Catalina, but
was ultimately caused by the early freeing of a git_repository instance,
by closing and reopening a repository whenever its location has changed
on the filesystem.

In order to support this change, window frame saving is suspended for
the reopened repository in order to avoid writing to the old location
of the repository.

Finally, the welcome window is no longer shown when all documents close
to prevent it from appearing during one of these re-openings. Clicking
on the GitUp icon in the dock will still show the welcome window. This
behaves similarly to Xcode.