Make diff view scale with font size
The font size matches the font size from user defaults directly.

The global variables associated with the diff view that have become properties on instances that update in `updateMetricsFromCurrentFontSize` when the user defaults change.

Constants have been replaced by functions.

The table view should always be reloaded when the font size changes to ensure the content size is correct.

The threshold for switching from unified to split diff is now specified as a multiple of the font size. Similar scaling is used for drawing the column separators for line numbers and positioning text. This ensures the layout is the same as GitUp was designed except for being scaled.