Text size setting (#620)
* Add font size user defaults key

Users should be able to control text size. The value in this user defaults key is the starting point for allowing any part of the GitUp UI to be scalable.

One limitation of storing the font size in the standard user defaults is that this is global state so an app using GitUpKit that wants to show two diff views at the same time would not be able to make them use different font sizes. I don’t think this use case is particularly important.

As a framework, GitUpKit should not register default values in user defaults, and it can’t guarantee that the app using GitUpKit registers a sensible default for this key. Therefore the font size should always be read from user defaults wrapped in a `GIFontSize()` function call, which ensures the value is greater than zero

The `GIFontSize` function was put in `GIAppKit.h/m` because that’s where other GitUpKit user defaults keys are defined. I didn’t think this fits in in `GIFunctions` because it’s only available in the Mac version of GitUpKit. Putting it in `GIFunctions` wrapped in `#if __GI_HAS_APPKIT__` would work too but makes less sense to me.

I thought it would be messy to do the `#ifdef __cplusplus` like in `GIFunctions` for this one function, so I used `FOUNDATION_EXPORT`. (My preferred way to do this would be to drop the `#ifdef __cplusplus` from `GIFunctions` and define a `GIEXPORT` used for every public function.)

Naming of the constant: I mirrored `GICommitMessageViewUserDefaultKey_ShowMargins` but removed the `CommitMessageView` part because this one is more general.

* Update MainMenu.xib

Modifications made automatically by Xcode 11.3 when opening first making a change to the file. There is nothing meaningful. This has been separated from the real change that is coming in the next commit.

* Make commit message view scale with font size

Must call `setNeedsDisplay:` otherwise the 50 and 72 character markers are sometimes not updated (such as when there is selected text).

The font size is 10% larger than the font size in user defaults, to match before where the font size here was 11 points compared to 10 in the diff view.

* Add slider in Preferences to change font size

Letting the user specify a scale but the application specify the typefaces and relative sizes seems like the best trade-off between allowing the user to see the app comfortably without asking them to be the designer. Also different views have use different fonts and relative sizes. Therefore, the user does not set an explicit typeface name or numerical size and instead a slider without value labels has been added to General Preferences to set the scale. This (deliberately) results in the exact size used hard to find.

I decided to store the most obvious thing in user defaults: the font size used in the diff view in points. This simplicity comes at the cost of making the slider in Preferences more complicated as a linear slider would not be very good: the difference between 10 and 11 points is much more important than the difference between 18 and 19 points. A value transformer is used to map between the discrete slider value and a non-linear array of font sizes that match the preset sizes in the Mac system font picker.

* Make diff view scale with font size

The font size matches the font size from user defaults directly.

The global variables associated with the diff view that have become properties on instances that update in `updateMetricsFromCurrentFontSize` when the user defaults change.

Constants have been replaced by functions.

The table view should always be reloaded when the font size changes to ensure the content size is correct.

The threshold for switching from unified to split diff is now specified as a multiple of the font size. Similar scaling is used for drawing the column separators for line numbers and positioning text. This ensures the layout is the same as GitUp was designed except for being scaled.