Add percentage to keep track of shown fractions
This property will be used to figure out how much to show of the old and
new image. A value of 0 will represent showing just the new image, while
a value of 1 will mean only the old image is shown.
Everything in between will move the divider in between accordingly.