[Feature] Image Diff Preview (#748)
* Introduce GIImageDiffView files

This view will be used very similarly to GIDiffView, as both are
variably sized diffing cell contents.

* Attach GIImageDiffView to GIDiffContentData

Again, similarly to the approach used for GIDiffView, we are attaching
this content view to the diff data, so it can later be integrated into
a cell.

* Show a GIImageDiffCellView for image diffs

Once again, this is mimicking GIDiffView and the associated
GITextDiffCellView. For now we merely use the previously introduced
imageDiffView as a way to identify whether an image cell should be
shown.

* Allow GIImageDiffView to request a height

This will later be necessary to request enough height to display images
at their full resolution (or a fraction of that depending on width
constraints).

* Inject GIImageDiffView into GIImageDiffCellView

Following the example of GIDiffView, we attach (and detach) the content
view to a designated cell.

* Inject the repository into each GIImageDiffView

To be able to export blobs of images later on, we'll need access to
the repository. This commit lays the necessary foundation for that step.

* Pass along the GCDiffDelta to GIImageDiffView

Adds another necessary resource to be able to display images later on.

* Add an image view for the current (new) image

* Update the frame of the current image view

* Update the current image when the delta is set

This preliminary solution only covers showing the absolute newest
version of the file and is therefore usually wrong.
It does however establish the connection between delta, repository and
the final image.

* Avoid unnecessary image updates

* Show the proper new file for all cases

This code is very similar to the viewDeltasInDiffTool utility method, as
we are fundamentally doing the same thing (exporting the image file
referenced by the SHA1 value and then displaying it).
The most noteworthy detail is that we are defaulting back to the
canonical path for untracked files.

* Add the second comparison image view

* Position the old image view alongside the new one

* Display the correct old comparison image

This works analogous to the previously added updateNewImage method.
The major difference is that we are looking at the oldFile and there
is no fallback if it does not exist.

* Add a method to calculate the combined image size

The goal is to make the GIImageDiffView a slide style view image
comparison tool, in which the two images are displayed in the same
area, but hidden or revealed by sliding a divider across them.
This method allows us to calculate how big this combined area needs
to be.

* Add a method to calculate the comparison frame

This method is designed to work within a given frame and will return
a rectangle that will provide the correct dimensions to display
both the old and new image in a shared area.

* Move old and new image into the shared area

Currently this will overlay the old image over the new one. However,
this is only setting up for the ability to partially hide both images.

* Make the image diff view ask for the proper height

Instead of asking for a somewhat random, large height to fit the image,
we can now ask for exactly the required height.
To do this, we try to fit the image at full resolution, but scale down
according to width constraints if necessary.

* Add percentage to keep track of shown fractions

This property will be used to figure out how much to show of the old and
new image. A value of 0 will represent showing just the new image, while
a value of 1 will mean only the old image is shown.
Everything in between will move the divider in between accordingly.

* Reset the percentage when delta is updated

* Partially hide the images according to percentage

By using masking layers, we can easily hide parts of the old and new
images. Originally, we considered drawing the images directly, but the
draw calls turned out to be very slow. This mask layer solution seems to
be about four times as fast.

* Update the percentage according to mouse input

By adding gesture recognizers to the view, we can use the entire image
area as a large input field for the user to select the split between
the old and new image.
Whenever we change the percentage value we need to trigger a redraw.

* Avoid implicit layer animation

By default, changing a CALayer's frame causes it to animate to the new
position. This prevents that behavior, as we need the change to happen
as fast as possible.

* Show the entire new image if there is no old one

* Add a transparency background layer

This adds the commonly used checkerboard transparency indicator.
The pattern had already been used by the GIMapViewController, so no new
assets were necessary.

* Update colors on every draw call

In order to reliably obey dark mode, it seems to be necessary to
reassign the colors constantly.
We tried to handle it via viewDidChangeEffectiveAppearance, but the
behavior was unreliable.
Since we couldn't notice any performance downsides to reassigning the
colors with every draw call, we went with this solution.

* Add indicator borders

Another common design element in slide style image comparison tools are
red and green borders to separate old and new images.
This is a simple implementation of that, which once again reassigns
the colors on every draw call to handle dark mode properly.

* Add a divider line

The final common design element of slider style image comparison tools.
To further indicate where the old and new images meet, we add a thin
line right between them.

* Infer file type based on UTI

The previous approach of initializing an NSImage to figure out whether a
file is an image was questionable when taking large binaries into account.
This new solution uses the file ending to determine the uniform type
identifier of a file, which can be used to check for image files.

* Limit maximum size of images in memory

To ensure fast performance and a reasonable memory impact even for very
large image files, we now create thumbnails of a limited size.

* Restrict image diffing to NSImage compatible types

* Separate image loading and cell size calculation

To ensure maximum performance we now offload loading the image into
memory to a separate thread.
As we still need the image size to lay out cells properly, we query
just this information before starting to load the full image.

* Show the entire deleted image