Update libgit2 fork to v1.9.0 (#2683)
* libgit2: bump submodule to updated fork

* libgit2: Remove tutorial comments from Package.swift

* libgit2 : use llhttp instead of http-client dependency

libgit2 changed which one it embeds.

* libgit2: update build excludes, build options

- Update source excludes (mostly, libgit2 source files were moved into new `libgit2, `util` and `cli` subfolders)
- Update compiler options to reflect changes in libgit2's CMake config

* libgit2: Use OpenSSH provider

This replaces the libssh2 SSH backend, and enables support of .ssh/config.

* libgit2: Silence most warnings

We're not interested in these when working on GitUp itself.

* libgit2: Hide incomplete deprecation warnings

These warnings are emitted *after* the actual libgit2 build, so these compiler options need to be in the project build settings, rather than the Third-Party package build settings.

This commit only adds -Wno-documentation-deprecated-sync to the each list; previously, we were using the default values, which are now explicit.

* Tests: Fix empty live repo tests by setting default branch name

These tests failed if you'd set the default branch name to e.g. main in your global Git config.

This builds on “Specify `initial_head` when initializing a repository for tests so the global git config doesn't make it fail (#988)” (aabbc8a).

* Add .editorconfig

* Don't crash when libgit2 writes to a closed pipe

This seems to happen whenever we pull from a remote with changes.