Rework mode and navigation item into single NSSegmentedControl
Tahoe really doesn't like styling or working with the custom views setup we had, this changes it to now be a single NSSegmentedControl that is programmatically converted between the mode and navigation setups.

With the macOS 11 bump, we also don't need a bunch of assets so we're removing those where they aren't needed.