Fix highlight range computation for non-ascii characters (#2742)
Fixes #2740

We were previously matching on the less specific 2-byte unicode point values, but this meant that something with with a 3 or 4 byte value (asian characters or emoji) would be first offset by 2 then loop incorrectly looking at the next character.

The fix is to look for the largest byte match first and then narrow down to the smaller bit matches.